Cover Flow: Build it yourself
Who needs external npm packages? I was about to use react-coverflow, but
first off it was a ridiculously large package, and it didn't even look
that good and used a bunch of dependencies that were unneeded.

So I built it myself! It's not perfect yet (my math is slightly off),
but it looks pretty dang close to the iPod cover flow.