Full-width Views: Set up container
This changes the way scrollIndex works (scrollIndex now lives within
each view's props) as well as how the "view container" is set up. There
is now a reusable "View Manager" in Toolbox that allows for multiple
view types to be displayed (or not) at the same time.

As of now, the two types of views are "full" and "split", and they
display/hide whenever a matching view type is pushed to the stack.